Pulled out all config structs (except p2p.PeerConfig) into config package
