+++ glide.lock
@@ -1,5 +1,5 @@
-hash: 6f8962f6ca0e25b8e43bc6e496bd46c9ff3a79dcf789578efeeaee2fffc39c6e
-updated: 2017-04-27T19:56:37.342860938-04:00
+hash: 1bdb98b332fc88df3659092ce571b74851f54465c11d91e273a10098dd4d8011
+updated: 2017-05-04T19:50:25.718540841+02:00
 imports:
 - name: github.com/btcsuite/btcd
   version: 4b348c1d33373d672edd83fc576892d0e46686d2
@@ -56,7 +56,7 @@ imports:
 - name: github.com/pelletier/go-toml
   version: 13d49d4606eb801b8f01ae542b4afc4c6ee3d84a
 - name: github.com/pkg/errors
-  version: 645ef00459ed84a119197bfb8d8205042c6df63d
+  version: bfd5150e4e41705ded2129ec33379de1cb90b513
 - name: github.com/pmezard/go-difflib
   version: d8ed2627bdf02c080bf22230dbb337003b7aba2d
   subpackages:
@@ -119,7 +119,7 @@ imports:
   subpackages:
   - term
 - name: github.com/tendermint/merkleeyes
-  version: ea4dd9c7b773435de26bf59fddf90afd43a07d67
+  version: 0fab643ccac1a3f93b90e0e2682a5d1b9d17f8c4
   subpackages:
   - app
   - client
@@ -129,6 +129,7 @@ imports:
   version: df250b69416a35a943a6e2a92118667e9ef031d4
   subpackages:
   - autofile
+  - cli
   - clist
   - common
   - db
