+++ glide.yaml
@@ -31,6 +31,7 @@ import:
   version: develop
   subpackages:
   - autofile
+  - cli
   - clist
   - common
   - db
